Added more keys to music keyboard
and cleaned up code that maps from keys to tones
12345678 maps to notes in selected scale, while
  W E       T  Y  U     O P
A  S  D  F  G  H  J  K  L  ;  '
maps to the black and white keys on a piano keyboard starting at C.
But the tones played are not necessarily those of a tempered 12-tone keyboard. Instead they are based on specific ratios to the specified First Tone, selecting the note of the Tones Per Octave tempered scale that is closest to the specific ratio for that key. For example, F is 4/3 and G is 3/2. Note that if Tones Per Octave is 12, the result is the standard tempered 12 tones.